requirements.txt
scipy
cython
matplotlib
scikit-image
tensorflow-gpu==1.15
keras==2.0.8
h5py==2.10.0
imgaug
IPython[all]
ffmpeg-python==0.2.0
future==0.18.2
numpy==1.19.1
opencv-python==4.4.0.42
pandas==1.1.1
Pillow==7.2.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
scipy==1.2.1